requirements.txt
selenium>=4.0.0
Flask>=2.0.0
Pillow>=8.0.0
opencv-python
pyyaml
requests
urllib3
numpy
